fix: remove unnecessary export from route handler
